+++ Source/libtess2/tesselator.h
@@ -143,8 +143,8 @@ typedef struct TESSalloc TESSalloc;
 // number of expected extra vertices.  
 struct TESSalloc
 {
-	void *(*memalloc)( void *userData, unsigned int size );
-	void *(*memrealloc)( void *userData, void* ptr, unsigned int size );
+	void *(*memalloc)( void *userData, size_t size );
+	void *(*memrealloc)( void *userData, void* ptr, size_t size );
 	void (*memfree)( void *userData, void *ptr );
 	void* userData;				// User data passed to the allocator functions.
 	int meshEdgeBucketSize;		// 512
